+++ src/pykeen/nn/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+"""PyKEEN internal "nn" module."""
+
+from . import init
+from .emb import Embedding, RepresentationModule
+
+__all__ = [
+    'Embedding',
+    'RepresentationModule',
+    'init',
+]
